+++ consensus/aquahash/algorithm_test.go
@@ -694,6 +694,7 @@ func TestConcurrentDiskCacheGeneration(t *testing.T) {
 		Extra:       []byte("www.bw.com"),
 		MixDigest:   common.HexToHash("0x3e140b0784516af5e5ec6730f2fb20cca22f32be399b9e4ad77d32541f798cd0"),
 		Nonce:       types.EncodeNonce(0xf400cd0006070c49),
+		Version:     types.H_KECCAK256,
 	})
 	// Simulate multiple processes sharing the same datadir
 	var pend sync.WaitGroup
